Move publisher stream implementation out of EventStream
