+++ src/main/scala/com/raquo/airstream/core/EventStream.scala
@@ -9,6 +9,7 @@ import com.raquo.airstream.debug.{DebuggableStream, Debugger, DebuggerStream}
 import com.raquo.airstream.distinct.DistinctStream
 import com.raquo.airstream.eventbus.EventBus
 import com.raquo.airstream.extensions._
+import com.raquo.airstream.javaflow.FlowPublisherStream
 import com.raquo.airstream.misc._
 import com.raquo.airstream.split.{SplittableOneStream, SplittableStream}
 import com.raquo.airstream.status.{AsyncStatusObservable, Status}
@@ -389,29 +390,12 @@ object EventStream {
     new JsPromiseStream[A](promise, emitOnce)
   }
 
+  /** Create a stream from a [[java.util.concurrent.Flow.Publisher]]
+    * - Use this to bring in events from other streaming libraries
+    * that can provide a `Flow.Publisher`, such as FS2 an Monix.
+    */
   def fromPublisher[A](publisher: Flow.Publisher[A], emitOnce: Boolean = false): EventStream[A] = {
-    var subscription: Flow.Subscription = null
-    fromCustomSource[A](
-      shouldStart = startIndex => if (emitOnce) startIndex == 1 else true,
-      start = (fireEvent, fireError, _, _) => {        
-        val subscriber = new Flow.Subscriber[A] {
-          def onNext(a: A) = fireEvent(a)
-          def onError(t: Throwable) = fireError(t)
-          def onComplete() = ()
-          def onSubscribe(s: Flow.Subscription) = {
-            s.request(Long.MaxValue)
-            subscription = s
-          }
-        }
-
-        publisher.subscribe(subscriber)
-      },
-      stop = _ => {
-        if (subscription ne null)
-          subscription.cancel()
-        subscription = null
-      }
-    )
+    FlowPublisherStream(publisher, emitOnce)
   }
 
   /** Easy helper for custom events. See [[CustomStreamSource]] for docs.
